Make physics constructors public for modularity
