+++ src/celeritas/ext/EmPhysicsList.cc
@@ -9,9 +9,9 @@
 #include <memory>
 
 #include "celeritas/Quantities.hh"
+#include "celeritas/g4/EmStandardPhysics.hh"
+#include "celeritas/g4/OpticalPhysics.hh"
 
-#include "detail/EmStandardPhysics.hh"
-#include "detail/OpticalPhysics.hh"
 #include "detail/PhysicsListUtils.hh"
 
 namespace celeritas
@@ -29,12 +29,12 @@ EmPhysicsList::EmPhysicsList(Options const& options)
         native_value_to<ClhepLen>(options.default_cutoff).value());
 
     // Celeritas-supported EM Physics
-    detail::emplace_physics<detail::EmStandardPhysics>(*this, options);
+    detail::emplace_physics<EmStandardPhysics>(*this, options);
 
     if (options.optical)
     {
         // Celeritas-supported Optical Physics
-        detail::emplace_physics<detail::OpticalPhysics>(*this, options.optical);
+        detail::emplace_physics<OpticalPhysics>(*this, options.optical);
     }
 }
