+++ src/celeritas/ext/FtfpBertPhysicsList.cc
@@ -17,10 +17,10 @@
 #include <G4ios.hh>
 
 #include "celeritas/Quantities.hh"
+#include "celeritas/g4/EmStandardPhysics.hh"
+#include "celeritas/g4/OpticalPhysics.hh"
 
-#include "detail/EmStandardPhysics.hh"
 #include "detail/MuHadEmStandardPhysics.hh"
-#include "detail/OpticalPhysics.hh"
 #include "detail/PhysicsListUtils.hh"
 
 namespace celeritas
@@ -38,17 +38,14 @@ FtfpBertPhysicsList::FtfpBertPhysicsList(Options const& options)
     this->SetDefaultCutValue(
         native_value_to<ClhepLen>(options.default_cutoff).value());
 
-    // Celeritas-supported EM physics
-    detail::emplace_physics<detail::EmStandardPhysics>(*this, options);
+    // Add celeritas EM physics plus additional mu/hadron
+    detail::emplace_physics<detail::MuHadEmStandardPhysics>(*this, options);
 
     if (options.optical)
     {
         // Celeritas-supported Optical Physics
-        detail::emplace_physics<detail::OpticalPhysics>(*this, options.optical);
+        detail::emplace_physics<OpticalPhysics>(*this, options.optical);
     }
-
-    // Hadron EM standard physics not supported in Celeritas
-    detail::emplace_physics<detail::MuHadEmStandardPhysics>(*this, verbosity);
 
     // TODO: Add a physics constructor equivalent to G4EmExtraPhysics
