+++ src/celeritas/g4/OpticalPhysics.hh
@@ -2,21 +2,19 @@
 // Copyright Celeritas contributors: see top-level COPYRIGHT file for details
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file celeritas/ext/detail/OpticalPhysics.hh
+//! \file celeritas/g4/OpticalPhysics.hh
 //---------------------------------------------------------------------------//
 #pragma once
 
 #include <G4VPhysicsConstructor.hh>
 
-#include "../GeantOpticalPhysicsOptions.hh"
+#include "celeritas/ext/GeantOpticalPhysicsOptions.hh"
 
 namespace celeritas
 {
-namespace detail
-{
 //---------------------------------------------------------------------------//
 /*!
- * Construct Celeritas-supported Optical physics.
+ * Construct Celeritas-supported optical physics.
  */
 class OpticalPhysics : public G4VPhysicsConstructor
 {
@@ -40,5 +38,4 @@ class OpticalPhysics : public G4VPhysicsConstructor
 };
 
 //---------------------------------------------------------------------------//
-}  // namespace detail
 }  // namespace celeritas
